More null checks for fedsrv crash on probe launcher assignment.

+++ src/Igc/probeigc.h
@@ -116,18 +116,20 @@ class CprobeIGC : public TmodelIGC<IprobeIGC>
                     if (oldFraction > 0.0f)
                     {
 						//Xynth Set flag in IGC ship to later trigger an achievement for destroying a probe
-						
 						if (launcher != nullptr && launcher->GetObjectType() == OT_ship)
 						{
 							ObjectID theID = launcher->GetObjectID();
 							IsideIGC * plSide = launcher->GetSide();
-							IshipIGC * pShip = plSide->GetShip(theID);
-							if (!((pside == launcher->GetSide()) || IsideIGC::AlliedSides(pside, launcher->GetSide())))
-								pShip->SetAchievementMask(c_achmProbeKill); //Xynth for enemy probe kill achievement
+
+							if (plSide != nullptr)
+							{
+								IshipIGC * pShip = plSide->GetShip(theID);
+
+								if (pShip != nullptr && pside != nullptr && !((pside == launcher->GetSide()) || IsideIGC::AlliedSides(pside, launcher->GetSide())))
+									pShip->SetAchievementMask(c_achmProbeKill); //Xynth for enemy probe kill achievement
+							}
 						}
 						
-					
-
                         GetMyMission()->GetIgcSite()->KillProbeEvent(this);
                         dr = c_drKilled;
                     }
